+++ lib/additionals/patches/project_patch.rb
@@ -15,7 +15,6 @@ module Additionals
       end
 
       module InstanceOverwriteMethods
-        # Used by Redmine >= 4.2
         def principals_by_role
           # includes = AdditionalsPlugin.active_hrm? ? [:roles, { principal: :hrm_user_type }] : %i[roles principal]
           includes = %i[principal member_roles roles]
@@ -25,21 +24,6 @@ module Additionals
 
               h[r] ||= []
               h[r] << m.principal
-            end
-            h
-          end
-        end
-
-        # Used by Redmine < 4.2
-        # this change take care of hidden roles and performance issues (includes for hrm, if installed)
-        def users_by_role
-          includes = AdditionalsPlugin.active_hrm? ? [:roles, { user: :hrm_user_type }] : %i[roles user]
-          members.includes(includes).each_with_object({}) do |m, h|
-            m.roles.each do |r|
-              next if r.hide && !User.current.allowed_to?(:show_hidden_roles_in_memberbox, project)
-
-              h[r] ||= []
-              h[r] << m.user
             end
             h
           end